Basic availability display functionality
