+++ frontend-react/src/components/calendar/Calendar.tsx
@@ -1,20 +1,11 @@
 import moment, { Moment } from 'moment';
-import React from 'react';
 import Day from './Day';
 
 interface Props {
     onDayClick: Function
-    highlightDays: number[]
-}
-
-const titleStyle = {
-    display: "flex",
-    flexDirection: "row" as "row",
-    justifyContent: "center"
-}
-
-const buttonStyle = {
-    margin: "10px 10px"
+    highlightDays: number[],
+    year: number,
+    month: number
 }
 
 const tableStyle = {
@@ -24,10 +15,11 @@ const tableStyle = {
 };
 
 export default function Calendar(props: Props) {
-    const [month, setMonth] = React.useState(1);
-    const [year, setYear] = React.useState(2022);
-  
-    let referenceDate: Moment = moment(new Date(year, month, 1));
+    const onDayClick = (day: number) => {
+        props.onDayClick(props.year, props.month, day)
+    }
+
+    let referenceDate: Moment = moment(new Date(props.year, props.month, 1));
     let firstDay: Number = Number(referenceDate.format("d"));
     let numDays: Number = referenceDate.daysInMonth();
 
@@ -52,7 +44,7 @@ export default function Calendar(props: Props) {
 
         week.push(
             <td key={i}>
-                <Day date={i} onClick={props.onDayClick} highlight={props.highlightDays.includes(i)} />
+                <Day date={i} onClick={onDayClick} highlight={props.highlightDays.includes(i)} />
             </td>
         );
     }
@@ -60,24 +52,8 @@ export default function Calendar(props: Props) {
     // Push the last, possibly incomplete, week into the calendar
     calendar.push(<tr key={0}>{week}</tr>);
 
-    const flip = (direction: number) => {
-        // + 12 because TS doesn't do well with negative moduluo
-        setMonth((month + direction + 12) % 12);
-        if (month + direction < 0) {
-            setYear(year - 1);
-        } else if (month + direction >= 12) {
-            setYear(year + 1);
-        } 
-    }
-
     return (
         <div>
-            <div style={titleStyle}>
-                <button style={buttonStyle} onClick={() => flip(-1)}>&lt;</button>
-                <p>{moment().month(month).format('MMMM')} {year}</p>
-                <button style={buttonStyle} onClick={() => flip(1)}>&gt;</button>
-            </div>
-            
             <table style={tableStyle}>
                 <thead>
                     <tr>
